Refactor session option to own file, and add option for mongodburi
