+++ .example.env
@@ -1,1 +1,7 @@
 PORT=8000
+
+# aviable option development/production
+# If you don't set this it will be development.
+NODE_ENV=development
+MONDOURI=yourmongodbURI
+
