+++ package.json
@@ -17,8 +17,9 @@
         "express": "^4.19.2",
         "express-session": "^1.18.0",
         "express-validator": "^7.1.0",
-        "mongodb": "^6.8.0",
-        "morgan": "^1.10.0"
+        "mongoose": "^8.5.1",
+        "morgan": "^1.10.0",
+        "passport": "^0.7.0"
     },
     "devDependencies": {
         "@eslint/js": "^8.57.0",
@@ -26,6 +27,7 @@
         "@types/express-session": "^1.18.0",
         "@types/morgan": "^1.9.9",
         "@types/node": "^20.14.10",
+        "@types/passport": "^1.0.16",
         "@typescript-eslint/eslint-plugin": "^7.16.0",
         "@typescript-eslint/parser": "^7.16.0",
         "eslint": "^8.57.0",
